// src/app/bizdevsyncbackend/models/activity.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Activity {
  Business_idBusiness?: number | null;
  Followup_idFollowup?: number | null;
  detail?: string | null;
  end_date?: string | null;
  idActivity?: number;
  is_archived?: boolean;
  last_action?: string | null;
  last_action_date?: string | null;
  next_action?: string | null;
  next_action_date?: string | null;
  priority?: 'CRITICAL' | 'IMPORTANT' | 'HIGH' | 'MEDIUM' | 'LOW';
  start_date?: string;
  status?: 'COMPLETED' | 'PENDING' | 'IN_PROGRESS' | 'NOT_STARTED' | 'WAITING_FEEDBACK';
  tags?: string | null;
  title?: string | null;
}